ml_service/requirements.txt
torch
torchvision
transformers
fastapi
uvicorn[standard]
Pillow
opencv-python-headless
mediapipe
numpy
pandas
scikit-learn
python-multipart
